arm64: KVM: make all pages non-executable by default
At the moment, we're killing the i-cache each time we map a page
in stage-2, as we have no idea if this page is used as a data or
code page.

In order to preserve the cache a bit more, mark all pages with the
XN flag. If we get an exec permission fault, invalidate the i-cache
and clear the XN bit.

We end up getting more faults, but we preserve the i-cache.

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>