GICv3: LPI/ITS support
GICv3 mandates the presence of an Interrupt Translation Service
(ITS) that behaves as a MSI controller.

The ITS allows a peripheral to generate a Locality-specific
Peripheral Interrupt (LPI) that is exposed by the GIC.

This patch adds support for LPIs as a separate IRQ domain,
and plugs the ITS into the MSI framework. Note that there is no
MSI-X support for the moment.

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>