kvmtool: AArch{32,64}: deal with duplicate cp15/sysreg accessors
The 3.14 kernel UAPI for KVM has now grown (or rather stolen!)
the cp15/sysreg accessors that kvmtool uses to deal with the
MPIDR access.

Test the existence of the macros so that kvmtool can be compiled
on both 3.13 and 3.14. Once the kvmtool repo is updated to 3.14,
the macros can be removed.

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>