PCI/MSI: Allow an msi_chip to be associated to an irq domain
The new MSI stacked domain has the interesting effect that it is
quite hard to use if the architecture doesn't provide its own
arch_setup_msi_irqs() function to override the default.

This inhibates the msi_chip infrastructure introduced in
0cbdcfcf427b (PCI: Introduce new MSI chip infrastructure), as it is
not possible to use it as an indirection between the core MSI code
and the MSI driver (useful when having multiple MSI controllers that
do not share the same requirements). The setup_irq and teardown_irq
are rendered useless, as their role is now taken over by the alloc/free
functions in the irq domain code.

This patch introduces an optionnal field to the msi_chip structure,
a pointer to an irq domain, describing the MSI domain associated
with this msi_chip. To be populated with msi_create_irq_domain.

If this field is non-NULL, then irq_domain_prepare_alloc_irqs()
called on this domain, bypassing the setup_irq callback. This allows
the MSI driver to use the domain stacking feature without mandating
core support in the architecture.

Tested on arm64 with the GICv3 ITS.

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>