irqdomain: Add prepare_alloc_irqs method
When performing the allocation of irqs within the domain, it is
sometimes necessary have a bus-specific operation that is going to
start the allocation process for a given device (device-specific
allocation/configuration) and in the end call a function similar
to pci_msi_domain_alloc_irqs().

For this purpose, introduce a prepare_alloc_irqs method to the
irq_domain structure, as well as a top-level helper that can
be called (irq_domain_prepare_alloc_irqs).

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>