irqchip/GICv3: Support FSL specific extention for shareable transactions
Freescale's integration of GIC500 (ARM's own GICv3 implementation) has
additional hardware enforcing shareable transaction, despite the GIC
reporting that the transactions are not shareable.

Introduce a new property (fsl,enforce-shareability) and a new set of
flags to deal with this feature. The property must be set at both
the GIC level (for the GICR_* registers) and at the ITS level (for
the GITS_* registers).

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>