arm64: KVM: VHE: Save guest sysreg minimal and auxiliary states
So far, we've considered the guest state to be monolithic. One
single bloc of registers. But in order to be able to run the host
kernel as early as possible, it would make more sense to split this
state in two:

- a "minimal" state: the set of register that allows the host's
  sysregs to be restored without loosing any state,
- an "auxiliary" state: the set of registers that complement the
  minimal state to form the complete sysreg state.

Obviously, the 32bit state is part of the auxiliary state. For
non-VHE, all the 64bit sysregs are part of the minimal state.
For VHE, this is limited to the sysregs shared between EL1 and EL2.

Once we have this in place, it becomes easy to defer saving the
guest state until it becomes convenient to do so.

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>