arm64: KVM: VHE: Use shadow HCR_EL2 for 32bit sysreg retore
In order to find out if we have to restore 32bit sysregs, we've
so far looked at HCR_EL2.RW. As we move non-essential state saving
towards the end of the world switch, we may end up in a situation
where HCR_EL2.RW has already been restored to the host's version.

Instead, we can test vcpu->arch.hcr_el2, which always contains
the right information.

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>