arm64: KVM: VHE: Save fpexc32_el2 as part of the FPSIMD context
Despite being an EL2 sysreg, FPEXC32_EL2 contains part of the
32bit VFP state, and thus is more of a FPSIMD register than an
actual sysreg.

Let's move its saving to the point where we actually save the
FPSIMD state. This has the benefit of making the rest of the 32bit
state independent of the state of the FP trapping, which means we'll
be able to move that state saving as well in a later patch.

We take this opportunity to make __fpsimd_enabled a static helper,
as it is only used in switch.c.

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>