KVM: arm/arm64: vgic-new: Don't BUG_ON on non-pending interrupt in sort
Having a BUG_ON in the interrupt sorting function when finding that
a non active interrupt is also not pending is not a great idea.
The pending state could have changed while we were not looking
(after all, we just locked the interrupt).

Instead, let's consider (enabled && pending) as a single state,
and make sure that the interrupt matching that state are moved
up the list.

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>